Refactored modules into es6 modules
